// src/report/report.types.ts
export type ReportType =
  | 'calories'
  | 'total_carbs'
  | 'total_fat'
  | 'protein'
  | 'trans'
  | 'monounsaturated_fat'
  | 'polyunsaturated_fat'
  | 'saturated_fat'
  | 'cholesterol'
  | 'sodium'
  | 'potassium'
  | 'fiber'
  | 'sugar'
  | 'vitamin_a'
  | 'vitamin_c'
  | 'calcium'
  | 'iron'
  | 'calories_burned'
  | 'mililiters';